requirements.txt
python-telegram-bot>=20.0
fastapi==0.104.1
uvicorn==0.24.0